выделил алгоритмы генерации в отдельный класс
