+++ WordCloudGenerator/Generator.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace WordCloudGenerator
 {
@@ -7,20 +8,9 @@ namespace WordCloudGenerator
     {
         private readonly Func<IEnumerable<WordFrequency>, IEnumerable<GraphicString>> algorithm;
 
-        public Generator(Func<IEnumerable<WordFrequency>, IEnumerable<GraphicString>> algorithm = null)
+        public Generator(Func<IEnumerable<WordFrequency>, IEnumerable<GraphicString>> algorithm)
         {
-            this.algorithm = algorithm ?? ExponentialAlgorithm;
-        }
-
-        private static IEnumerable<GraphicString> ExponentialAlgorithm(IEnumerable<WordFrequency> wordFrequencies)
-        {
-            var fontSize = 400f;
-
-            foreach (var wordFrequency in wordFrequencies)
-            {
-                yield return new GraphicString(wordFrequency.Word, Math.Max(75, fontSize));
-                fontSize /= 1.1f;
-            }
+            this.algorithm = algorithm;
         }
 
         public IEnumerable<GraphicString> CalculateFontSizeForWords(IEnumerable<WordFrequency> wordFrequencies)
